pro_architecture/requirements-pro.txt
setuptools>=65.0.0
wheel

flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

pinecone-client==5.0.1
neo4j==5.14.0

openai>=1.30.0
cohere==5.6.3

psutil==5.9.6
python-dotenv==1.0.1
